13_julia_fp: send location in float, rather than int
This avoids an int-to-float cast that is currently unimplemented

+++ test/unit_tests/13_julia_fp/kernel.h
@@ -11,5 +11,5 @@
 #include <stdio.h>
 
 extern "C" {
-void func(int32_t *a, int32_t *b);
+void func(float *a, float *b);
 }
